Footer Mockup RD1
Using USDA site (https://www.usda.gov/) for inspiration in knowing that it uses the standards and that the FS is an agency of the USDA. Please note, the green bar will be a duplicate of the navigation. Its there as a placeholder for now until we know what nav items will be included.