Permit Paths Mockups RD1
Here are two options for the visual design of the permit path screens. Both include a mobile and desktop version. Both include photography to show real life scenarios of people interacting in the forests. The green horizontal bars are nav placeholders. Each permit type panel includes the permit type name, a description of the permit type and a button to go to the next step. The "I'm not sure" permit type panel has been styled differently in each option to help differentiate it as it is not an official permit type. One option uses photography within each permit type panel and the other uses a large background photo similar to how the background photo is used on the existing Forest Service website. There is some concern that using a large background photo may cause some bandwidth issues, however, we can create tests to get a better picture of those issues.