Admin User Mockups
Posting both the applications list and application detail screen mockups. In the list mockup, have altered the UI to more of a card layout for a few different reasons. 1. More visually appealing 2. Easier to scan and probably most important 3. Easier viewing/transition to mobile.  The user also has the ability to sort app cards via the sort dropdown or search for specific app cards. Each card gives the basic info (dates, event name and location) and allows the user to select and move to the detail view. Within the detail view, user is presented with all application information. They can then take three actions, Accept the application, if they need additional info Hold the application which then visually flags it or Return the application. The Edit button allows the admin to adjust details if needed. The thought behind the button bar at the bottom is that it would be a sticky footer, always being at the bottom of the user's browser therefore always available to Accept, Hold or Return an app