Temp Outfitter Form Nav Mockup RD1
Incorporating a nav to the temp outfitter form. The thought is that it would always be viewable (sticky nav) and will indicate where the user is in the form (see the Operating Plan step styling in comparison to others meaning the user is currently on that step). If all fields in a category pass validation, a checkmark would show up next to them. If there are errors in that step, an error label would appear.