Application Revision History Mockups RD3
Updated mockups. Summary of mockups:

application-revision-history-RD1-01.jpg - An example of revision
history in a pending application detail.

application-revision-history-RD1-02.jpg -
An example of revision history in an accepted application detail.

application-revision-history-RD1-03.jpg -
An example of revision history in a rejected application detail.

application-revision-history-RD1-04.jpg -
An example of revision history in a cancelled application detail.