Fix error running migrations on staging
- Throw an error in a user-provided key does not exist
- Only get the user provided database service if the DATABASE_URL does not
  exist.

In cloud.gov, DATABASE_URL is provided and no "database" user-provided service
needs to exist. In that case, we should check for the presence of the
environment variable and only fallback to reading the user-provided service if
it doesn't exist. If neither exist, the app will crash as expected since there's
not point in operating without a database.